requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
pyjwt==2.8.0

# AI Services
openai==1.3.7
google-cloud-vision==3.4.5
elevenlabs==0.2.26

# Image Processing
pillow==10.1.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Email Validation
email-validator==2.1.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging
python-json-logger==2.0.7